+++ src/indexer.h
@@ -132,7 +132,9 @@ inline void Reflect(Writer& visitor, IndexFuncRef& value) {
 
   if (value.is_implicit)
     s += "~";
-  if (value.id.id == -1) {  // id.id is unsigned, special case 0 value
+
+  // id.id is unsigned, special case 0 value
+  if (static_cast<ssize_t>(value.id.id) == -1) {
     s += "-1";
   } else {
     s += std::to_string(value.id.id);
